Add files via upload
This is the Python 2.7.13 version of the plant time checking program requested in Item 62 of the Python course. I was unable to get the tzinfo class module (or any other time-related mod other than Datetime) to do correct conversions, and pytz would not load into my Python 2 using pip, so I was forced to use deduction to solve this problem. London is UTC time (usually), Portland is local time, and New York is Portland + 3 hours.