Add files via upload
This is the Python Item 63 exercise, which transfers files from Folder_A to Folder_B with a single  button press. Given that it was only a one button GUI, I decided not to make it as modular as the Phone Book program. GitHub would not let me upload the Folder_B directory since it was empty. The program was designed to run from the E:\ drive, given that was where the folders originally were stored.