Add files via upload
This is the Python Item 64 exercise, which checks files to see if they are over 24 hours old, then commits them to an outbound folder called "Outgoing" from source folder "Incoming". Note that this program will look for files and folders on the E: Drive, so locations may have to be changed, but all else should work.